+++ index.d.ts
@@ -1,4 +1,5 @@
 export * from "./core";
 export * from "./ga4";
+export * from "./runner";
 export * from "./telegram";
 export * from "./vercel";