+++ index.js
@@ -1,6 +1,7 @@
 module.exports = {
   ...require("./core"),
   ...require("./ga4"),
+  ...require("./runner"),
   ...require("./telegram"),
   ...require("./vercel")
 };